// src/redux/profile/profile.types.ts
enum ProfileTypes {
	FETCH_PROFILE_INFO_START = "FETCH_PROFILE_INFO_START",
	FETCH_PROFILE_INFO_SUCCESS = "FETCH_PROFILE_INFO_SUCCESS",
	FETCH_PROFILE_INCIDENTS_START = "FETCH_PROFILE_INCIDENTS_START",
	FETCH_PROFILE_INCIDENTS_SUCCESS = "FETCH_PROFILE_INCIDENTS_SUCCESS",
	ADD_FRIEND_START = "ADD_FRIEND_START",
	ADD_FRIEND_SUCCESS = "ADD_FRIEND_SUCCESS",
	REMOVE_FRIEND_START = "REMOVE_FRIEND_START",
	REMOVE_FRIEND_SUCCESS = "REMOVE_FRIEND_SUCCESS",
	PROFILE_ERROR = "PROFILE_ERROR",
	FETCH_PROFILE_FRIENDS_START = "FETCH_PROFILE_FRIENDS_START",
	FETCH_PROFILE_FRIENDS_SUCCESS = "FETCH_PROFILE_FRIENDS_SUCCESS",
	SET_IS_FRIEND = "SET_IS_FRIEND",
}

export default ProfileTypes;